Moving complex coefficient code into separate files
